requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.6.1
python-dotenv>=1.0.0
openai>=1.12.0
chromadb>=0.4.22
duckdb>=0.9.2
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
httpx>=0.26.0
PyJWT>=2.8.0
python-dateutil>=2.8.2
aiosmtplib>=2.0.2
jinja2>=3.1.3 